refactor: move formatting utilities to shared package
- Create generic formatters in packages/shared/src/utils/formatters.ts
- Includes formatNumber, formatDuration, formatTimestamp, formatRelativeTime, escapeHtml
- Update dashboard to re-export from shared package
- Export formatters from shared package index
- Maintain backward compatibility for dashboard
